refactor: extract splitSections util
